- fix looping sounds - clientsetup (use current mfc, 64bit build, speaker + disableaudio options)

+++ src/game/client/application/SwgClientSetup/src/win32/MessageBox2.h
@@ -38,7 +38,7 @@ public:
 	// ClassWizard generated virtual function overrides
 	//{{AFX_VIRTUAL(MessageBox2)
 	public:
-	virtual int DoModal();
+	virtual intptr_t DoModal();
 	protected:
 	virtual void DoDataExchange(CDataExchange* pDX);    // DDX/DDV support
 	//}}AFX_VIRTUAL
